Move RRulePeroperties' tests out
