+++ rrule-debugger/Cargo.toml
@@ -11,7 +11,7 @@ publish = false
 rrule = { version = "0.7.0", path = "../rrule" }
 chrono = "0.4.19"
 chrono-tz = "0.6.1"
-clap = "3.1.15"
+clap = { version = "3.1.15", features = ["derive"] }
 rrule-afl-fuzz = { version = "0.1.0", path = "../rrule-afl-fuzz" }
 log = "0.4.16"
 yansi = "0.5.1"
